Fix circular dependencies
It was mostly splitting up interfaces from their class files
