+++ src/braintree-manager/braintree-interfaces.ts
@@ -1,4 +1,3 @@
-import { PaymentProvidersInterface } from './payment-providers';
 import { DonationRequest } from '../models/request-models/donation-request';
 import { DonationResponse } from '../models/response-models/donation-response';
 import { SuccessResponse } from '../models/response-models/success-models/success-response';
@@ -7,6 +6,7 @@ import { BillingInfo } from '../models/common/billing-info';
 import { CustomerInfo } from '../models/common/customer-info';
 import { DonationPaymentInfo } from '../models/donation-info/donation-payment-info';
 import { PaymentProvider } from '../models/common/payment-provider-name';
+import { PaymentProvidersInterface } from './payment-providers-interface';
 
 /**
  * The BraintreeManager is the main entrypoint for much of the common braintree functionality.
