+++ src/braintree-manager/payment-providers/apple-pay/apple-pay-interface.ts
@@ -1,0 +1,13 @@
+import { PromisedSingleton } from '@internetarchive/promised-singleton';
+import { DonationPaymentInfo } from '../../../models/donation-info/donation-payment-info';
+import { ApplePaySessionDataSourceInterface } from './apple-pay-session-datasource-interface';
+
+export interface ApplePayHandlerInterface {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  instance: PromisedSingleton<any | undefined>;
+  isAvailable(): Promise<boolean>;
+  createPaymentRequest(
+    e: Event,
+    donationInfo: DonationPaymentInfo,
+  ): Promise<ApplePaySessionDataSourceInterface>;
+}
