+++ src/braintree-manager/payment-providers/google-pay-interface.ts
@@ -1,0 +1,8 @@
+import { PromisedSingleton } from '@internetarchive/promised-singleton';
+
+export interface GooglePayHandlerInterface {
+  paymentsClient: google.payments.api.PaymentsClient;
+  instance: PromisedSingleton<braintree.GooglePayment>;
+
+  isBrowserSupported(): Promise<boolean>;
+}
