+++ src/braintree-manager/payment-providers/venmo-interface.ts
@@ -1,0 +1,8 @@
+import { PromisedSingleton } from '@internetarchive/promised-singleton';
+
+export interface VenmoHandlerInterface {
+  instance: PromisedSingleton<braintree.Venmo>;
+
+  isBrowserSupported(): Promise<boolean>;
+  startPayment(): Promise<braintree.VenmoTokenizePayload>;
+}
