+++ test/mocks/payment-providers/individual-providers/mock-creditcard-handler.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { CreditCardHandlerInterface } from '../../../../src/braintree-manager/payment-providers/credit-card/credit-card';
 import { PromisedSingleton } from '@internetarchive/promised-singleton';
 import { HostedFieldName } from '../../../../src/braintree-manager/payment-providers/credit-card/hosted-field-container';
 import { MockHostedFieldsClient } from '../../payment-clients/mock-hostedfields-client';
 import { mockHostedFieldTokenizePayload } from '../../payment-clients/mock-hostedfieldtokenizepayload';
+import { CreditCardHandlerInterface } from '../../../../src/braintree-manager/payment-providers/credit-card/credit-card-interface';
 
 export class MockCreditCardHandler implements CreditCardHandlerInterface {
   instance: PromisedSingleton<braintree.HostedFields> = new PromisedSingleton<
