+++ test/mocks/payment-providers/individual-providers/mock-paypal-handler.ts
@@ -1,9 +1,9 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { PayPalHandlerInterface } from '../../../../src/braintree-manager/payment-providers/paypal/paypal';
 import { PromisedSingleton } from '@internetarchive/promised-singleton';
 import { MockPayPalClient } from '../../payment-clients/mock-paypal-client';
 import { DonationPaymentInfo } from '../../../../src/models/donation-info/donation-payment-info';
 import { PayPalButtonDataSourceInterface } from '../../../../src/braintree-manager/payment-providers/paypal/paypal-button-datasource';
+import { PayPalHandlerInterface } from '../../../../src/braintree-manager/payment-providers/paypal/paypal-interface';
 
 export class MockPayPalHandler implements PayPalHandlerInterface {
   instance: PromisedSingleton<braintree.PayPalCheckout | undefined> = new PromisedSingleton<
